[CONFIGURATION] * BUG: Try to detect correctly all X11 linking paramters.
[64 BITS]
* NEW: Start to remove all ambiguous datatypes.

[DEVELOPMENT ENVIRONMENT]
* NEW: The gambas2 project was renamed to gambas3.
* BUG: The completion and signature font sizes are now the default one.
* NEW: Add support for Basque language.

[WIKI CGI SCRIPT]
* BUG: The symbol pages work again, and inheritance between class belonging
  to different components now is taken into account correctly.

[SCRIPTER]
* NEW: The gbs2 project was renamed to gbs3.


git-svn-id: svn://localhost/gambas/trunk@914 867c0c6c-44f3-4631-809d-bfa615b0a4ec

+++ main/share/gb_file_share.h
@@ -37,7 +37,7 @@ typedef
   struct {
     short type;
     short mode;
-    long long size;
+    int64_t size;
     int atime;
     int mtime;
     int ctime;
@@ -124,7 +124,7 @@ PUBLIC void FILE_recursive_dir(const char *dir, void (*found)(const char *), voi
 
 PUBLIC void FILE_make_path_dir(const char *path);
 
-PUBLIC long long FILE_free(const char *path);
+PUBLIC int64_t FILE_free(const char *path);
 
 #else
