+++ main/share/gb_pcode_temp.h
@@ -47,9 +47,9 @@ PUBLIC short PCODE_dump(FILE *out, short addr, PCODE *code)
   int j;
   unsigned short op;
   unsigned short digit;
-  long value;
+  int value;
   #ifdef PROJECT_COMP
-  long index;
+  int index;
   TABLE *table;
   bool trans;
   #endif
@@ -255,7 +255,7 @@ PUBLIC short PCODE_dump(FILE *out, short addr, PCODE *code)
           break;
 
         case C_PUSH_LONG:
-          value = *((long *)&code[1]);
+          value = *((int *)&code[1]);
           fprintf(out, "PUSH INTEGER %ld", value);
           break;
